kernel/services/field_chat/requirements.txt
# SAHOOL Field Chat Dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0
httpx>=0.25.0
nats-py>=2.6.0
tortoise-orm>=0.20.0
asyncpg>=0.29.0
aerich>=0.7.2
websockets>=12.0
python-dateutil>=2.8.2
aiosqlite>=0.19.0,<0.20.0  # For SQLite testing
